Fixed MSVC warning
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@27320 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/sip/sdp.cxx
@@ -443,7 +443,7 @@ bool SDPMediaFormat::PostDecode(const OpalMediaFormatList & mediaFormats, unsign
       m_mediaFormat.AddOption(new OpalMediaOptionString(SDPBandwidthPrefix + r->first, false, r->second), true);
   }
 
-  if (bandwidth > 1000 && bandwidth < m_mediaFormat.GetOptionInteger(OpalMediaFormat::MaxBitRateOption())) {
+  if (bandwidth > 1000 && bandwidth < (unsigned)m_mediaFormat.GetOptionInteger(OpalMediaFormat::MaxBitRateOption())) {
     PTRACE(4, "SDP\tAdjusting format \"" << m_mediaFormat << "\" bandwidth to " << bandwidth);
     m_mediaFormat.SetOptionInteger(OpalMediaFormat::MaxBitRateOption(), bandwidth);
   }
